imgtool: Create simple signing tool
The imgtool is a simple image signing tool based on the libraries in the
`newt` tool, but appropriate for images that live outside of a Mynewt
project.

This initial version is able to generate, extract public keys from, and
sign ECDSA P-256 keys.  The signing tool is also able to add an image
trailer to mark this image for upgrade.