doc: design: clear up definition of an image trailer
The design document says that image trailers can be located in
scratch. That's not true: the contents in scratch are just a subset of
the complete image trailer.

Fix this up by making it clear that image trailers are at the end of
image areas, and sometimes some of their data is copied into and out
of scratch.

Signed-off-by: Marti Bolivar <marti.bolivar@linaro.org>