doc: design: fixes and improvements for swap status definition
The swap status field definition claims image data are swapped one
sector at a time, but that's not true: image data are actually swapped
around in increments equal to the size of the scratch area, which
can be multiple sectors in length.

Its contents are also misleading in claiming that it's a series of
single byte records, which ignores padding when min-write-size > 1.

It also assumes the reader knows how images are swapped, but that's
not explained until later on in the document.

Finally, the paragraph is in a list of "definitions" of the image
trailer's fields, but it doesn't actually define the contents; it's
just a high level description of what the field is for.

Fix all of these issues by re-working this paragraph.

Signed-off-by: Marti Bolivar <marti.bolivar@linaro.org>