Rearrange docs for gh-pages
Github pages allows the documentation to be in the master branch in a
'docs' directory to be rendered as the main site (mcuboot.com).  Rename
this directory, and pull in the documentation files from the old
gh-pages branch.

The main index.md page does not link to the rest of the docs yet, and
that change can be made in a future patch.

Signed-off-by: David Brown <david.brown@linaro.org>