zephyr: add Kconfig options for CONF_xxx
The Zephyr CMakeLists.txt contains some configuration options that are
outside of Kconfig.

The standard way to configure Zephyr builds is with Kconfig, so it
would be better if MCUboot did not have a parallel configuration
mechanism.

Add equivalent Kconfig options to those which are defined in
CMakeLists.txt. Subsequent patches will use them.

Signed-off-by: Marti Bolivar <marti@opensourcefoundries.com>