ext: mbedtls: Change platform settings to avoid warnings
calloc and free are actually not used in our usage of mbedtls. In order
to get rid of warnings when building it, enable the
MBEDTLS_PLATFORM_MEMORY and MBEDTLS_PLATFORM_NO_STD_FUNCTIONS macros which
will then rely on the system to provide them for us or fail at build
time.

Signed-off-by: Carles Cufi <carles.cufi@nordicsemi.no>