sim: CBOR/SUIT support
Add libraries to the simulator to support generation of SUIT manifests.
This has its own CBOR implementation because the existing Rust crates
for CBOR are focused on marshaling data, and not encoding to match a
specific format.

The SUIT support is based on the 3rd draft of the IETF WG manifest
document.  Since this is supporting a draft RFC, there may need to be
changes as newer drafts or a final version come out.

Signed-off-by: David Brown <david.brown@linaro.org>