Boot: Enable multi-image boot in overwrite only mode
This patch adds the capability to handle multiple firmware images, to
update them independently. Currently this is supported only in
overwrite-only mode when MCUBOOT_OVERWRITE_ONLY is set.
It introduces the boot_set_active_image() function which can be used
to select the active image and to invalidate the keys when switching
between images in that case if the secondary image slot is encrypted.

It separates the completion of aborted image swap operations and the
update of images even more as these should be happening at different
stages of the boot process according to the design proposal of
the multiple image support:
https://github.com/JuulLabs-OSS/mcuboot/pull/317 .

Change-Id: I7eb5f632298bb08c805bfaee0359703b2ae19e9d
Signed-off-by: David Vincze <david.vincze@arm.com>
Signed-off-by: David Brown <david.brown@linaro.org>